07-real-world-app/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langgraph==0.0.20
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
httpx==0.25.2
aiofiles==23.2.1
jinja2==3.1.2 